Module Ft.Synthesize

type t = {

bold : bool;

oblique : bool;

}

val get : [ `Ft ] Font_face.t -> t

get font returns the synthesized information.

val set : ?bold:bool -> ?oblique:bool -> [ `Ft ] Font_face.t -> unit

set font synthesize different glyphs from a base font, which is useful if you lack those glyphs from a true bold or oblique font.

val unset : ?bold:bool -> ?oblique:bool -> [ `Ft ] Font_face.t -> unit

unset font undo what set did.